update final version

+++ src/include/encode/bfilter/bfilter_encode.h
@@ -21,7 +21,7 @@
 
 class BFilterEncode : public Encode {
 public:
-    BFilterEncode(uint32_t best_hash_nums, uint32_t hash_size) : best_hash_nums_(best_hash_nums), hash_size_(hash_size),
+    BFilterEncode(uint32_t best_hash_nums, uint64_t hash_size) : best_hash_nums_(best_hash_nums), hash_size_(hash_size),
                                                                  hash_param1_(HASH_PARAM1), hash_param2_(HASH_PARAM2),
                                                                  Encode() {}
 
@@ -32,15 +32,15 @@ public:
      *          hash param
      *  @return :  return (key*hash_param1+key2*hash_param2)%|bitset|
      * */
-    inline uint32_t Hash(uint32_t key1, uint32_t key2, uint32_t hash_param1, uint32_t hash_param2) {
-        return (key1 * hash_param1 + key2 * hash_param2) % hash_size_;
+    inline uint64_t Hash(uint32_t key1, uint32_t key2, uint32_t hash_param1, uint32_t hash_param2) {
+        return ((uint64_t)key1 * hash_param1 + (uint64_t)key2 * hash_param2) % hash_size_;
     };
 
 protected:
     uint32_t best_hash_nums_;
     std::vector<uint32_t> hash_param1_;
     std::vector<uint32_t> hash_param2_;
-    uint32_t hash_size_;
+    uint64_t hash_size_;
 };
 
 #endif //VEND_BLOOM_FILTER_ENCODE_H
